final version

+++ Game/Source/EnemyAir.cpp
@@ -16,6 +16,7 @@
 
 EnemyAir::EnemyAir(pugi::xml_node paras) : Entity(EntityType::ENEMYAIR)
 {
+	parameters = paras;
 	name.Create("EnemyAir");
 }
 
@@ -24,28 +25,34 @@ EnemyAir::~EnemyAir() {
 }
 
 bool EnemyAir::Awake() {
-
+	InitSpawn(parameters);
 	return true;
 }
 
 bool EnemyAir::Start() {
 
-	health = 2;
+	health = maxHealth = 2;
 	pbody->body->SetGravityScale(0.0f);
 	return true;
 }
 
 bool EnemyAir::Update()
 {
-	if (health <= 0) {
-		pbody->body->DestroyFixture(pbody->body->GetFixtureList());
-		app->entityManager->DestroyEntity(this);
-	}
+	if (app->scene->gamePaused == false) {
+		if (health <= 0) {
+			pugi::xml_node item;
+			app->scene->CreateItem(item, Item::ItemType::HEALTH, position.x, position.y);
 
-	FindPath();
-	Move();
-	RenderEntity();
-	UpdateAnim();
+			pbody->body->DestroyFixture(pbody->body->GetFixtureList());
+			app->entityManager->DestroyEntity(this);
+		}
+
+		FindPath();
+		Move();
+		RenderEntity();
+		UpdateAnim();
+	}
+	
 	return true;
 }
 
@@ -208,9 +215,12 @@ void EnemyAir::RenderEntity()
 
 	if (!app->scene->playerptr->deadTextureOn)
 	{
+		HealthBar();
 		currentAnimation->Update();
 		SDL_Rect rect1 = currentAnimation->GetCurrentFrame();
 		app->render->DrawTexture(texture, position.x, position.y, &rect1);
+
+		
 	}
 }
 
@@ -230,4 +240,10 @@ void EnemyAir::UpdateAnim()
 	{
 		currentAnimation = &idle;
 	}
+}
+
+void EnemyAir::HealthBar() {
+	healthPerc = ((100 / maxHealth) * health);
+	SDL_Rect healthbar = { (position.x - 5), position.y - 10, healthPerc * 0.5, 5 };
+	app->render->DrawRectangle(healthbar, 255, 0, 0, 255, true, true);
 }
